pyproject.toml
[project]
name = "deep-research"
version = "1.0.0"
description = "Grounded multi-agent deep research system — cited, scored, vaulted."
requires-python = ">=3.10"
dependencies = [
  "anthropic>=0.40.0",
  "openai>=1.50.0",
  "google-generativeai>=0.8.0",
  "tavily-python>=0.5.0",
  "requests>=2.31.0",
  "arxiv>=2.1.0",
  "psycopg2-binary>=2.9.9",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "tenacity>=8.3.0",
  "tiktoken>=0.7.0",
  "networkx>=3.3",
  "aiohttp>=3.9.0",
]

[project.scripts]
deepresearch = "deep_research.cli.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
